requirements.txt
absl-py>=0.12.0
cachetools>=5.0.0
chex>=0.0.7
clu>=0.0.6
jax>=0.2.25
flax>=0.3.6
ml-collections==0.1.0
numpy>=1.19.5
optax>=0.1.0
scipy>=1.4.0
tensorflow-cpu>=2.4.0  # Using tensorflow-cpu to have all GPU memory for JAX.
tensorflow-datasets>=4.0.1
